Try to prevent "cannot commit" errors by checking for pending earlier.

+++ Sqliteman/sqliteman/alterviewdialog.cpp
@@ -14,10 +14,12 @@ for which a new license (GPL+exception) is in place.
 #include "database.h"
 
 
-AlterViewDialog::AlterViewDialog(const QString & name, const QString & schema, QWidget * parent)
+AlterViewDialog::AlterViewDialog(const QString & name, const QString & schema,
+								 LiteManWindow * parent)
 	: QDialog(parent),
 	update(false)
 {
+	creator = parent;
 	ui.setupUi(this);
 	ui.databaseCombo->addItem(schema);
 	ui.nameEdit->setText(name);
@@ -45,30 +47,33 @@ AlterViewDialog::AlterViewDialog(const QString & name, const QString & schema, Q
 
 void AlterViewDialog::createButton_clicked()
 {
-	update = true;
-	ui.resultEdit->clear();
-	QString sql(QString("DROP VIEW \"%1\".\"%2\"")
-			.arg(ui.databaseCombo->currentText())
-			.arg(ui.nameEdit->text()));
-	QSqlQuery dropQuery(sql, QSqlDatabase::database(SESSION_NAME));
-	if (dropQuery.lastError().isValid())
+	if (creator && creator->checkForPending())
 	{
-		ui.resultEdit->insertPlainText(tr("Error while altering view (drop phase): %1.\n\n%2").arg(dropQuery.lastError().text()).arg(sql));
-		ui.resultEdit->moveCursor(QTextCursor::Start);
-	}
+		update = true;
+		ui.resultEdit->clear();
+		QString sql(QString("DROP VIEW \"%1\".\"%2\"")
+				.arg(ui.databaseCombo->currentText())
+				.arg(ui.nameEdit->text()));
+		QSqlQuery dropQuery(sql, QSqlDatabase::database(SESSION_NAME));
+		if (dropQuery.lastError().isValid())
+		{
+			ui.resultEdit->insertPlainText(tr("Error while altering view (drop phase): %1.\n\n%2").arg(dropQuery.lastError().text()).arg(sql));
+			ui.resultEdit->moveCursor(QTextCursor::Start);
+		}
 
-	sql = QString("CREATE VIEW \"%1\".\"%2\" AS\n%3;")
-			.arg(ui.databaseCombo->currentText())
-			.arg(ui.nameEdit->text())
-			.arg(ui.sqlEdit->text());
-	QSqlQuery query(sql, QSqlDatabase::database(SESSION_NAME));
-	
-	if(query.lastError().isValid())
-	{
-		ui.resultEdit->insertPlainText(tr("Error while altering view: %1.\n\n%2").arg(query.lastError().text()).arg(sql));
+		sql = QString("CREATE VIEW \"%1\".\"%2\" AS\n%3;")
+				.arg(ui.databaseCombo->currentText())
+				.arg(ui.nameEdit->text())
+				.arg(ui.sqlEdit->text());
+		QSqlQuery query(sql, QSqlDatabase::database(SESSION_NAME));
+		
+		if(query.lastError().isValid())
+		{
+			ui.resultEdit->insertPlainText(tr("Error while altering view: %1.\n\n%2").arg(query.lastError().text()).arg(sql));
+			ui.resultEdit->insertPlainText("\n");
+			return;
+		}
+		ui.resultEdit->insertPlainText(tr("View altered successfully"));
 		ui.resultEdit->insertPlainText("\n");
-		return;
 	}
-	ui.resultEdit->insertPlainText(tr("View altered successfully"));
-	ui.resultEdit->insertPlainText("\n");
 }
